Speed up initial shell loading
The initial zshell loading was taking over a second to load. I updated
my configuration in order to speed this up. The biggest things that I
did involve eliminating external shell calls (especially to slow
languages like ruby). Updating the way that files are sourced, removing
rbenv since it adds 200ms to load times just on its own, and optimizing
when compinit is called. Unfortunately compinit is just damn slow and
often has to audit and dump a new compinit file out which can take ~60ms
or more. In the worst case it'll take almost 2 seconds to load a shell
just because of how slow compinit is. I'm going to be looking for a
solution to this later. But for now its very tolerable.